chore: moved app from routes folder
